test(agent_templates): move template tool tests to agent_templates server
Moves search_agent_templates and get_agent_template tests out of agent_sidekick_context.test.ts into a new agent_templates.test.ts. Exports TOOLS from index.ts to allow direct handler testing.
