+++ front/lib/api/actions/servers/agent_templates/index.ts
@@ -62,4 +62,5 @@ function createServer(
   return server;
 }
 
+export { TOOLS };
 export default createServer;
